+++ examples/vue2/src/components/layout/header/title-logo.vue
@@ -1,9 +1,0 @@
-<template>
-  <WizHeading fontSize="sm" color="gray.900" padding="0 1rem">
-    Wiz UI Playground
-  </WizHeading>
-</template>
-
-<script setup lang="ts">
-import { WizHeading } from "@wizleap-inc/wiz-ui";
-</script>