+++ packages/wiz-ui-next/src/components/base/inputs/timepicker/time-picker.vue
@@ -43,6 +43,7 @@
                   :class="[
                     timePickerSelectorOptionStyle,
                     timePickerSelectorOptionItemStyle,
+                    timePickerCursorStyle[timePickerCursor],
                     option === selectedHour &&
                       timePickerSelectorOptionItemSelectedStyle,
                     timePickerSelectorOptionItemColorStyle[
@@ -79,6 +80,7 @@
                   :class="[
                     timePickerSelectorOptionStyle,
                     timePickerSelectorOptionItemStyle,
+                    timePickerCursorStyle[timePickerCursor],
                     option === selectedMinute &&
                       timePickerSelectorOptionItemSelectedStyle,
                     timePickerSelectorOptionItemColorStyle[