+++ packages/wiz-ui-react/src/components/base/inputs/search-selector/components/search-selector.tsx
@@ -42,6 +42,7 @@ type Props = BaseProps & {
   addable?: boolean;
   error?: boolean;
   isDirectionFixed?: boolean;
+  showExLabel?: boolean;
   dropdownMaxHeight?: string;
   onChangeValues: (values: number[]) => void;
   onCreate?: (label: string) => void;
@@ -61,6 +62,7 @@ const SearchSelector: FC<Props> = ({
   addable = false,
   error,
   isDirectionFixed = false,
+  showExLabel = false,
   dropdownMaxHeight,
   onChangeValues,
   onCreate,
@@ -103,6 +105,7 @@ const SearchSelector: FC<Props> = ({
           option: {
             label: option.label,
             value: option.value,
+            exLabel: option.exLabel,
             onClick: () => {
               setSearchText("");
               if (multiSelectable) {
@@ -248,6 +251,9 @@ const SearchSelector: FC<Props> = ({
               >
                 <span className={styles.selectBoxInnerBoxSelectedLabelStyle}>
                   {selectedOption.label}
+                  {showExLabel &&
+                    selectedOption.exLabel &&
+                    ` (${selectedOption.exLabel})`}
                 </span>
                 <button
                   type="button"