+++ packages/wiz-ui-react/src/components/base/inputs/time-picker/components/time-picker.tsx
@@ -146,6 +146,7 @@ const TimePicker: FC<Props> = ({
                     key={"hh" + option}
                     className={clsx([
                       styles.timePickerSelectorOptionStyle,
+                      styles.timePickerCursorStyle[timePickerCursor],
                       option !== time?.hour &&
                         styles.timePickerSelectorOptionItemStyle,
                       option === time?.hour &&
@@ -191,6 +192,7 @@ const TimePicker: FC<Props> = ({
                     key={"mm" + option}
                     className={clsx([
                       styles.timePickerSelectorOptionStyle,
+                      styles.timePickerCursorStyle[timePickerCursor],
                       option !== time?.minute &&
                         styles.timePickerSelectorOptionItemStyle,
                       option === time?.minute &&