+++ packages/wiz-ui/src/components/base/anchor/index.ts
@@ -1,1 +1,0 @@
-export { default as WizAnchor } from "./anchor.vue";