+++ packages/wiz-ui/src/components/base/buttons/toggle/index.ts
@@ -1,1 +1,0 @@
-export { default as WizToggleButton } from "./toggle.vue";