+++ packages/wiz-ui/src/components/base/divider/index.ts
@@ -1,1 +1,0 @@
-export { default as WizDivider } from "./divider.vue";