+++ packages/wiz-ui/src/components/base/dropdown/index.ts
@@ -1,2 +1,0 @@
-export { default as WizDropdown } from "./dropdown.vue";
-export { default as WizDropdownItem } from "./dropdown-item.vue";