+++ packages/wiz-ui/src/components/base/graphs/compare/index.ts
@@ -1,1 +1,0 @@
-export { default as WizCompareGraph } from "./compare.vue";