+++ packages/wiz-ui/src/components/base/graphs/meter/index.ts
@@ -1,1 +1,0 @@
-export { default as WizMeterGraph } from "./meter.vue";