+++ packages/wiz-ui/src/components/base/graphs/transition/index.ts
@@ -1,1 +1,0 @@
-export { default as WizTransitionGraph } from "./transition.vue";