+++ packages/wiz-ui/src/components/base/heading/index.ts
@@ -1,1 +1,0 @@
-export { default as WizHeading } from "./heading.vue";