+++ packages/wiz-ui/src/components/base/inputs/search-selector/index.ts
@@ -1,1 +1,0 @@
-export { default as WizSearchSelector } from "./search-selector.vue";