+++ packages/wiz-ui/src/components/base/tables/index.ts
@@ -1,3 +1,0 @@
-export * from "./card";
-export * from "./flat";
-export * from "./unstyled";