+++ packages/wiz-ui/src/types/storybook-vue-router.d.ts
@@ -1,5 +1,0 @@
-declare module "storybook-vue-router" {
-  export default function storyRouterDecorator(
-    routes: import("vue-router").RouteConfig[]
-  ): import("@storybook/vue-vite").StoryFnVueReturnType;
-}